build: rewrite build, start scripts
