+++ config/webpack.config.dev.js
@@ -50,7 +50,8 @@ var plugins = [
   // Generates an `index.html` file with the <script> injected.
   new HtmlWebpackPlugin({
     inject: true,
-    template: paths.appHtml
+    template: paths.appHtml,
+    favicon: paths.appIco
   }),
   // Makes some environment variables available in index.html.
   // The public URL is available as %PUBLIC_URL% in index.html, e.g.:
