+++ package.json
@@ -26,15 +26,11 @@
   "scripts": {
     "init": "npm install commitizen -g && npm i",
     "commit": "git-cz",
-    "pre-release": "npm run release -- --prerelease rc",
     "release": "standard-version",
-    "release-major": "npm run release -- --release-as major",
-    "release-minor": "npm run release -- --release-as minor",
-    "release-patch": "npm run release -- --release-as patch",
     "test-old": "./node_modules/.bin/mocha --compilers coffee:coffee-script,js:babel/register --require coffee-script/register",
     "test": "jest",
-    "debug": "node --inspect scripts/start_n.js",
-    "start": "node scripts/start_n.js",
+    "debug": "node --inspect scripts/start.js",
+    "start": "node scripts/start.js",
     "start-prod": "PROD=true node server.js",
     "stylelint": "stylelint '**/*.scss'",
     "stylelint:fix": "stylelint '**/*.scss' --fix",
@@ -71,12 +67,11 @@
     "indexeddbshim": "^2.0.4",
     "intl": "^1.1.0",
     "jdenticon": "git+https://github.com/cryptonomex/jdenticon.git",
-    "jest-cli": "^23.4.0",
+    "jest-cli": "^24.9.0",
     "js-sha256": "^0.2.3",
     "jsdom": "^11.11.0",
     "lodash": "^4.17.15",
     "lzma": "2.1.6",
-    "machina": "^2.0.0",
     "moment": "^2.17.1",
     "moment-timezone": "^0.5.11",
     "node-fetch": "^1.3.1",
@@ -92,7 +87,7 @@
     "react-date-picker": "^5.3.28",
     "react-dom": "^15.2.0",
     "react-foundation-apps": "git+https://github.com/valzav/react-foundation-apps.git",
-    "react-highcharts": "^8.4.2",
+    "react-highcharts": "^16.0.2",
     "react-hot-loader": "^4.12.12",
     "react-interpolate-component": "^0.9.0",
     "react-intl": "^2.0.1",
@@ -121,16 +116,14 @@
     "reselect": "^2.5.4",
     "rx": "^2.5.2",
     "secure-random": "1.1.1",
-    "svgo": "^0.5.1",
+    "svgo": "^1.3.0",
     "tcomb": "2.5.2",
-    "websocket": "^1.0.18",
     "whatwg-fetch": "^0.11.0"
   },
   "devDependencies": {
     "@commitlint/cli": "^8.1.0",
     "@commitlint/config-conventional": "^8.1.0",
-    "autoprefixer": "6.7.2",
-    "babel": "^5.8.29",
+    "babel": "^6.23.0",
     "babel-core": "^6.26.3",
     "babel-eslint": "^8.2.5",
     "babel-jest": "^24.1.0",
@@ -174,17 +167,15 @@
     "file-loader": "^4.2.0",
     "html-webpack-plugin": "^4.0.0-beta.5",
     "husky": "^3.0.4",
-    "image-webpack-loader": "^3.0.0",
+    "image-webpack-loader": "^5.0.0",
     "inquirer": "^7.0.0",
     "json-loader": "^0.5.7",
     "lint-staged": "^8.1.6",
     "mocha": "^6.2.0",
-    "mocha-traceur": "~2.1.0",
     "node-fetch": "^1.3.1",
     "node-libs-browser": "~0.5.2",
     "node-localstorage": "~0.3.4",
     "node-sass": "^4.12.0",
-    "npm-run": "^4.1.0",
     "opn": "^5.4.0",
     "peerplaysjs-lib": "^0.6.0",
     "postcss-import": "^12.0.1",
@@ -209,7 +200,7 @@
     "svgo-loader": "^1.1.0",
     "traceur": "0.0.110",
     "uglify-es": "git://github.com/mishoo/UglifyJS2.git#harmony",
-    "url-loader": "^0.5.7",
+    "url-loader": "^2.1.0",
     "webpack": "^4.39.3",
     "webpack-dev-server": "^3.8.0",
     "webpack-manifest-plugin": "^2.0.4",
